backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
scikit-learn
pandas
numpy
joblib
matplotlib
python-dotenv
alembic
